// src/app/model/user.ts
export interface User {
  login: string;
  id: number;
  avatar_url: string;
  url: string;
  html_url: string;
  repos_url: string;

  name: string;
  email: string;
  company: string;
  location: string;
  bio: string;
  public_repos: number;
  public_gist: number;
  followers: number;
  following: number;

}